added simple spotify connect POC
