+++ client/src/Food.js
@@ -1,0 +1,9 @@
+import React from 'react';
+
+const Food = () => (
+    <div>
+      <span>Food</span>
+    </div>
+);
+
+export default Food;
