+++ client/src/Home.js
@@ -1,0 +1,9 @@
+import React from 'react';
+
+const Home = () => (
+    <div>
+      <span>Home</span>
+    </div>
+);
+
+export default Home;
